Generate system Nexus WIT with protoc plugin
