--- cmd/protoc-gen-nexus-rpc-yaml/go.mod
@@ -1,12 +1,0 @@
-module github.com/temporalio/api/cmd/protoc-gen-nexus-rpc-yaml
-
-go 1.25.4
-
-require (
-	google.golang.org/protobuf v1.36.1
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require github.com/nexus-rpc/nexus-proto-annotations v0.0.0-20260330194009-e558d6edaf84
-
-require github.com/google/go-cmp v0.6.0 // indirect
