+++ cmd/protoc-gen-system-nexus-wit/go.mod
@@ -1,0 +1,8 @@
+module github.com/temporalio/api/cmd/protoc-gen-system-nexus-wit
+
+go 1.25.4
+
+require (
+	github.com/nexus-rpc/nexus-proto-annotations v0.0.0-20260330194009-e558d6edaf84
+	google.golang.org/protobuf v1.36.1
+)
