+++ cmd/protoc-gen-system-nexus-wit/main.go
@@ -1,0 +1,11 @@
+// protoc-gen-system-nexus-wit generates nexus/temporal-system.wit from proto service
+// methods annotated with option (nexusannotations.v1.operation).tags = "exposed".
+package main
+
+import "google.golang.org/protobuf/compiler/protogen"
+
+func main() {
+	protogen.Options{}.Run(func(gen *protogen.Plugin) error {
+		return generate(gen)
+	})
+}
