+++ nexus/temporal-system.wit
@@ -4,7 +4,7 @@ world system {
   export workflow-service;
 }
 
-/// @nexus.endpoint "temporal-system"
+/// @nexus.endpoint "__temporal_system"
 interface workflow-service {
   use nexus:temporal-types/model@1.0.0.{
     duration,
